feat

+++ Transactions/Resources/Show/PlanResource.cs
@@ -5,6 +5,6 @@ public class PlanResource
     public int Id { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public long Fee { get; set; }
+    public double Fee { get; set; }
     public SupplierResource Supplier { get; set; }
 }
